feat(consensus): finalize DKG speculatively (#1477)
Changes DKG finalization to work on notarized blocks.

Tempo wants to finalize its DKG ceremony and include the outcome in the
boundary block of each epoch. With this patch, the application actor
requests the DKG outcome based on the boundary block's parent digest.
The DKG manager then decides whether it can already construct the
outcome of the DKG ceremony (because it has already finalized the
pre-to-boundary block), or if it has to first fetch notarized blocks and
construct the outcome based on those.

Before this patch, the DKG actor would only construct the DKG outcome
upon finalization of the pre-to-boundary block. But because simplex
gives no guarantees about observing direct finalizations of all blocks,
this could lead to chain halts in certain edge cases (network
instability comes to mind).

With this patch the node no longer persists the dealings it finds in
finalized blocks. Instead, on restart it backfills them from the marshal
actor.

Closes #924